requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.4
coloredlogs==15.0.1
dataclasses-json==0.5.9
decorator==5.1.1
distlib==0.3.6
ffmpeg-python==0.2.0
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
humanfriendly==10.0
identify==2.5.24
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
langchain==0.0.228
langchainplus-sdk==0.0.20
lazy_loader==0.3
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.8.0
numba==0.57.1
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.8.1
pooch==1.7.0
pre-commit==3.3.3
protobuf==4.23.4
pyarrow==12.0.1
pycryptodomex==3.18.0
pydantic==1.10.11
pydeck==0.8.0
pydub==0.25.1
Pygments==2.15.1
PyMatting==1.1.8
Pympler==1.0.1
python-dateutil==2.8.2
python-ulid==1.1.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.29.1
rembg==2.0.49
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
scikit-image==0.21.0
scipy==1.9.3
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.18
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
tifffile==2023.7.4
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
virtualenv==20.23.1
websockets==11.0.3
yarl==1.9.2
youtube-transcript-api==0.6.1
yt-dlp==2023.7.6
zipp==3.16.0
